feat: change cci logo on footer (#1351)
