+++ apps/web/src/assets/scss/setting/_typography.scss
@@ -11,5 +11,6 @@ $font-styles: (
 );
 
 $text-decorations: (
-  underline
+  underline,
+  none
 );
