+++ apps/web/src/components/identification/details/TeeProfileDetails.vue
@@ -5,7 +5,7 @@
   >
     <TeeDsfrButton
       size="small"
-      class="fr-btn--tertiary-no-outline fr-text-left fr-p-0 fr-text--white fr-btn-bg fr-text--sm fr-text--underline"
+      class="fr-btn--tertiary-no-outline fr-text-left fr-p-0 fr-text--white fr-btn-bg fr-text--sm fr-text--decoration-underline"
       @click="openSiretStep"
     >
       <span class="fr-pr-2v fr-icon-arrow-left-line" /><span>je renseigne mon SIRET</span>
